LPS-196591 Move LockedLayoutsGroupConfiguration to layout-api
